+++ dynamic-app/src/androidTest/java/com/okta/idx/android/cucumber/definitions/LoginDefinitions.kt
@@ -23,7 +23,7 @@ import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withHint
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
-import com.okta.idx.android.dynamic.R
+import com.okta.idx.android.auth.R
 import com.okta.idx.android.infrastructure.ACTION_BUTTON_VIEW
 import com.okta.idx.android.infrastructure.ERROR_TEXT_VIEW
 import com.okta.idx.android.infrastructure.EndToEndCredentials
