+++ dynamic-app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.okta.idx.android.dynamic">
+    package="com.okta.idx.android.auth">
 
     <application
         android:name=".SampleApplication"
