+++ dynamic-app/src/main/java/com/okta/idx/android/auth/LaunchFragment.kt
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.okta.idx.android.dynamic
+package com.okta.idx.android.auth
 
 import android.os.Bundle
 import android.view.View
 import androidx.navigation.fragment.findNavController
-import com.okta.idx.android.dynamic.databinding.FragmentLaunchBinding
+import com.okta.idx.android.auth.databinding.FragmentLaunchBinding
 import com.okta.idx.android.util.BaseFragment
 
 internal class LaunchFragment : BaseFragment<FragmentLaunchBinding>(
