+++ dynamic-app/src/main/java/com/okta/idx/android/auth/dynamic/DynamicAuthField.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.okta.idx.android.dynamic.auth
+package com.okta.idx.android.auth.dynamic
 
 import android.content.Context
 import android.graphics.Bitmap
